@@ -4,57 +4,57 @@ publish_to: 'none'
 version: 3.4.2+41
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.10.0 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ">=3.16.0 <4.0.0"
 
 dependencies:
   auto_size_text: ^3.0.0
-  cached_network_image: ^3.2.3
-  collection: ^1.17.1
+  cached_network_image: ^3.3.0
+  collection: ^1.18.0
   crypto: ^3.0.3
   datetime_picker_formfield: ^2.0.1
-  fl_chart: ^0.63.0
+  fl_chart: ^0.64.0
   flutter:
     sdk: flutter
   flutter_acrcloud: ^0.2.1
-  flutter_cache_manager: ^3.3.0
-  flutter_local_notifications: ^14.1.1
+  flutter_cache_manager: ^3.3.1
+  flutter_local_notifications: ^16.1.0
   flutter_mpmediaplayer: ^0.3.0
   flutter_web_auth: ^0.5.0
   gallery_saver: ^2.3.2
   html: ^0.15.4
-  http: ^1.0.0
-  in_app_review: ^2.0.6
+  http: ^1.1.0
+  in_app_review: ^2.0.8
   intl: ^0.18.1
   json_annotation: ^4.8.1
   modal_bottom_sheet: ^2.1.2
-  octo_image: ^1.0.2
-  package_info_plus: ^4.0.2
-  path_provider: ^2.0.15
+  octo_image: ^2.0.0
+  package_info_plus: ^4.2.0
+  path_provider: ^2.1.1
   pkce: ^1.1.0+1
   pool: ^1.5.1
-  quick_actions: ^1.0.5
+  quick_actions: ^1.0.6
   rxdart: ^0.27.7
-  share_plus: ^7.0.2
-  shared_preferences: ^2.1.2
-  sqflite: ^2.2.8+4
+  share_plus: ^7.2.1
+  shared_preferences: ^2.2.2
   uni_links: ^0.5.1
-  universal_html: ^2.2.3
+  sqflite: ^2.3.0
+  universal_html: ^2.2.4
   universal_io: ^2.2.2
-  url_launcher: ^6.1.11
+  url_launcher: ^6.2.1
   visibility_detector: ^0.4.0+2
   wakelock: ^0.6.2
-  workmanager: ^0.5.1
+  workmanager: ^0.5.2
 
 dev_dependencies:
-  build_runner: ^2.4.5
+  build_runner: ^2.4.6
   flutter_launcher_icons: ^0.13.1
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
-  json_serializable: ^6.7.0
+  json_serializable: ^6.7.1
 
 dependency_overrides:
   flutter_launcher_icons:
